Move EnvVarRepresenter to shared module and some minor fixes
