+++ .gitignore
@@ -204,6 +204,9 @@ api/api-shared-v1/target/
 api/api-shared-v4/config/
 api/api-shared-v4/out/
 api/api-shared-v4/target/
+api/api-shared-v2/config/
+api/api-shared-v2/out/
+api/api-shared-v2/target/
 api/api-shared-v6/config/
 api/api-shared-v6/logs/
 api/api-shared-v6/out/
