+++ api/api-environments-v2/src/main/java/com/thoughtworks/go/apiv2/environments/EnvironmentsControllerV2.java
@@ -25,11 +25,11 @@ import com.thoughtworks.go.api.spring.ApiAuthenticationHelper;
 import com.thoughtworks.go.api.util.GsonTransformer;
 import com.thoughtworks.go.api.util.HaltApiMessages;
 import com.thoughtworks.go.api.util.MessageJson;
-import com.thoughtworks.go.apiv2.environments.exception.InvalidGoCipherTextRuntimeException;
 import com.thoughtworks.go.apiv2.environments.model.PatchEnvironmentRequest;
 import com.thoughtworks.go.apiv2.environments.representers.EnvironmentRepresenter;
 import com.thoughtworks.go.apiv2.environments.representers.EnvironmentsRepresenter;
 import com.thoughtworks.go.apiv2.environments.representers.PatchEnvironmentRequestRepresenter;
+import com.thoughtworks.go.apiv2.shared.exception.InvalidGoCipherTextRuntimeException;
 import com.thoughtworks.go.config.BasicEnvironmentConfig;
 import com.thoughtworks.go.config.EnvironmentConfig;
 import com.thoughtworks.go.config.exceptions.EntityType;
