+++ api/api-environments-v2/src/main/java/com/thoughtworks/go/apiv2/environments/representers/PatchEnvironmentRequestRepresenter.java
@@ -18,6 +18,7 @@ package com.thoughtworks.go.apiv2.environments.representers;
 
 import com.thoughtworks.go.api.representers.JsonReader;
 import com.thoughtworks.go.apiv2.environments.model.PatchEnvironmentRequest;
+import com.thoughtworks.go.apiv2.shared.representers.EnvironmentVariableRepresenter;
 import com.thoughtworks.go.config.EnvironmentVariableConfig;
 
 import java.util.ArrayList;
