+++ api/api-shared-v2/build.gradle
@@ -14,11 +14,13 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.apiv2.environments.exception;
+apply plugin: 'jacoco'
+apply plugin: 'groovy'
 
-public class InvalidGoCipherTextRuntimeException extends RuntimeException {
+dependencies {
+  compile project(':api:api-base')
+  testCompile project(path: ':api:api-base', configuration: 'testOutput')
 
-    public InvalidGoCipherTextRuntimeException(String message, Throwable throwable) {
-        super(message, throwable);
-    }
+  testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: project.versions.junit5
+  testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: project.versions.junit5
 }
