+++ api/api-shared-v2/src/test/groovy/com/thoughtworks/go/apiv2/shared/representers/EnvironmentVariableRepresenterTest.groovy
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.apiv2.environments.representers
+package com.thoughtworks.go.apiv2.shared.representers
 
 import com.thoughtworks.go.api.util.GsonTransformer
 import com.thoughtworks.go.config.EnvironmentVariableConfig
 import com.thoughtworks.go.security.GoCipher
+import org.hamcrest.MatcherAssert
 import org.junit.jupiter.api.Test
 import spark.HaltException
 
@@ -132,6 +133,6 @@ class EnvironmentVariableRepresenterTest {
       EnvironmentVariableRepresenter.fromJSON(jsonReader)
     })
 
-    assertThat(haltException.body(), is("{\n  \"message\" : \"Environment variable must contain either 'value' or 'encrypted_value'\"\n}"))
+    MatcherAssert.assertThat(haltException.body(), is("{\n  \"message\" : \"Environment variable must contain either 'value' or 'encrypted_value'\"\n}"))
   }
 }
