+++ settings.gradle
@@ -63,6 +63,7 @@ include ':api:api-server-health-messages-v1'
 include ':api:api-server-health-v1'
 include ':api:api-server-maintenance-mode-v1'
 include ':api:api-shared-v1'
+include ':api:api-shared-v2'
 include ':api:api-shared-v4'
 include ':api:api-shared-v6'
 include ':api:api-stage-operations-v1'
